core: replace native NNS with non-native NNS in core tests
